+++ app/helpers/alchemy/base_helper.rb
@@ -2,11 +2,6 @@
 
 module Alchemy
   module BaseHelper
-    def _t(key, *args)
-      ActiveSupport::Deprecation.warn("Alchemys `_t` method is deprecated! Use `Alchemy.t` instead.", caller.unshift)
-      Alchemy.t(key, *args)
-    end
-
     # An alias for truncate.
     # Left here for downwards compatibilty.
     def shorten(text, length)