+++ app/helpers/alchemy/url_helper.rb
@@ -16,36 +16,6 @@ module Alchemy
       alchemy.show_page_url(show_page_path_params(page, optional_params))
     end
 
-    # This helper returns a path to picture for use inside a image_tag helper.
-    #
-    # Any additional options are passed to the url_helper, so you can add arguments to your url.
-    #
-    # Example:
-    #
-    #   <%= image_tag show_alchemy_picture_path(picture, size: '320x200', format: :png) %>
-    #
-    # @deprecated Use Alchemy::Picture#url instead
-    #
-    def show_alchemy_picture_path(picture, optional_params = {})
-      ActiveSupport::Deprecation.warn("`show_alchemy_picture_path` helper is deprecated and will be removed in Alchemy 4.0. Please use the `picture.url` method instead.")
-      picture.url(optional_params)
-    end
-
-    # This helper returns an url to picture for use inside a image_tag helper.
-    #
-    # Any additional options are passed to the url_helper, so you can add arguments to your url.
-    #
-    # Example:
-    #
-    #   <%= image_tag show_alchemy_picture_url(picture, size: '320x200', format: :png) %>
-    #
-    # @deprecated Use Alchemy::Picture#url instead
-    #
-    def show_alchemy_picture_url(picture, optional_params = {})
-      ActiveSupport::Deprecation.warn("`show_alchemy_picture_url` helper is deprecated and will be removed in Alchemy 4.0. Please use the `picture.url` method instead.")
-      picture.url(optional_params)
-    end
-
     # Returns the correct params-hash for passing to show_page_path
     def show_page_path_params(page, optional_params = {})
       raise ArgumentError, 'Page is nil' if page.nil?